コードの修正2回目
